# docs/api-spec.yml
...
openapi: 3.1.0
info:
  title: Archivo Desclasificado — Editorial Service
  version: 0.1.0
paths:
...
  /chapters/{chapter_id}/video/generate:
    post:
      tags:
      - video-generation
      summary: Generate Video
      description: 'Start generating videos; returns 202 immediately with pending
        rows.


        Composition runs in the background — poll GET /chapters/{id}/video for

        status. Defaults to every platform of the chapter in Spanish only;

        generating both languages doubles narration cost. Returns 409 when the

        script is unapproved or a pending/generated video already exists for that

        platform and language.'
      operationId: generate_video_chapters__chapter_id__video_generate_post
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/VideoGenerateRequest'
              - type: 'null'
              title: Request
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoGenerationResponse'
                title: Response Generate Video Chapters  Chapter Id  Video Generate
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chapters/{chapter_id}/video:
    get:
      tags:
      - video-generation
      summary: Get Videos
      description: Soft-deleted attempts are excluded here; GET /chapters/{id}/audit
        keeps them.
      operationId: get_videos_chapters__chapter_id__video_get
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoGenerationResponse'
                title: Response Get Videos Chapters  Chapter Id  Video Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chapters/{chapter_id}/video/retry:
    post:
      tags:
      - video-generation
      summary: Retry Video
      description: 'Retry failed generations, reusing audio and visuals already on
        disk.


        Retries every failed attempt for the chapter, or one when

        video_generation_id is given. Returns 409 when nothing has failed.'
      operationId: retry_video_chapters__chapter_id__video_retry_post
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/VideoRetryRequest'
              - type: 'null'
              title: Request
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoGenerationResponse'
                title: Response Retry Video Chapters  Chapter Id  Video Retry Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /videos/{video_id}:
    delete:
      tags:
      - video-management
      summary: Delete One Video
      description: 'Delete a video''s files and mark the record deleted.


        Soft delete: the row is retained so the retry lineage the audit trail

        reports stays intact. Shared narration and canonical subtitles are never

        removed, and a per-platform file is kept when another live record still

        points at it — the response says which and why.'
      operationId: delete_one_video_videos__video_id__delete
      parameters:
      - name: video_id
        in: path
        required: true
        schema:
          type: string
          title: Video Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDeletionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VideoDeletionResponse:
      properties:
        video_generation_id:
          type: string
          title: Video Generation Id
        video_file_deleted:
          type: boolean
          title: Video File Deleted
        subtitle_file_deleted:
          type: boolean
          title: Subtitle File Deleted
        video_file_retained_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Video File Retained Reason
        subtitle_file_retained_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle File Retained Reason
      type: object
      required:
      - video_generation_id
      - video_file_deleted
      - subtitle_file_deleted
      - video_file_retained_reason
      - subtitle_file_retained_reason
      title: VideoDeletionResponse
    VideoGenerateRequest:
      properties:
        platforms:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Platforms
          description: Platform names; defaults to all of the chapter's.
        languages:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Languages
          description: Language codes; defaults to Spanish only.
      type: object
      title: VideoGenerateRequest
    VideoGenerationResponse:
      properties:
        id:
          type: string
          title: Id
        platform_version_id:
          type: string
          title: Platform Version Id
        platform:
          type: string
          title: Platform
        language:
          type: string
          title: Language
        status:
          type: string
          title: Status
        video_file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Video File Path
        subtitle_file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle File Path
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        retry_of_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Retry Of Id
        created_at:
          type: string
          format: date-time
          title: Created At
        generated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Generated At
        size_mb:
          anyOf:
          - type: number
          - type: 'null'
          title: Size Mb
      type: object
      required:
      - id
      - platform_version_id
      - platform
      - language
      - status
      - video_file_path
      - subtitle_file_path
      - error_message
      - retry_of_id
      - created_at
      - generated_at
      - size_mb
      title: VideoGenerationResponse
...
    VideoRetryRequest:
      properties:
        video_generation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Video Generation Id
          description: Retry one attempt; defaults to every failed one.
      type: object
      title: VideoRetryRequest
...
servers:
- url: http://localhost:8000
  description: Local development server