# src/editorial/presentation/routers/video_generation.py
"""
Video generation endpoints — specs/video-generation-from-script/spec.md.

Generation is asynchronous by requirement: the POST persists PENDING rows and
returns their ids immediately, then composition runs in a FastAPI background
task while the admin panel polls GET for status. design.md Decision 7 rejected
a task queue (Celery/RQ) as operational overhead for an editorial workflow, so
the background runs in-process — meaning a server restart mid-composition
leaves rows PENDING, which the retry endpoint exists to clear up.
"""
from datetime import datetime
from typing import Optional

from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException
from pydantic import BaseModel, Field
from sqlalchemy.orm import Session, sessionmaker

from src.editorial.application.script_approval import (
    ChapterNotFoundError,
    is_script_approved,
)
from src.editorial.application.video_management import file_size_mb
from src.editorial.application.video_generation_use_case import (
    SUPPORTED_LANGUAGES,
    GenerationAlreadyExistsError,
    ScriptNotApprovedError,
    VideoGenerationUseCase,
)
from src.editorial.infrastructure.persistence.models import (
    Chapter,
    VideoGeneration,
    VideoGenerationStatus,
)
from src.editorial.infrastructure.persistence.session import (
    get_session,
    get_session_factory,
)
from src.editorial.presentation.dependencies import get_video_generation_use_case

router = APIRouter(tags=["video-generation"])

# Spanish is the authored language; English is a translation an editor should
# opt into after reviewing the translated subtitles, so it is not generated by
# default — a default of both would double TTS spend on every click.
DEFAULT_LANGUAGES = ("es",)


class VideoGenerateRequest(BaseModel):
    platforms: Optional[list[str]] = Field(
        default=None, description="Platform names; defaults to all of the chapter's."
    )
    languages: Optional[list[str]] = Field(
        default=None, description="Language codes; defaults to Spanish only."
    )


class VideoRetryRequest(BaseModel):
    video_generation_id: Optional[str] = Field(
        default=None, description="Retry one attempt; defaults to every failed one."
    )


class VideoGenerationResponse(BaseModel):
    id: str
    platform_version_id: str
    platform: str
    language: str
    status: str
    video_file_path: Optional[str]
    subtitle_file_path: Optional[str]
    error_message: Optional[str]
    retry_of_id: Optional[str]
    created_at: datetime
    generated_at: Optional[datetime]
    size_mb: Optional[float]


def _to_response(record: VideoGeneration) -> VideoGenerationResponse:
    return VideoGenerationResponse(
        id=record.id,
        platform_version_id=record.platform_version_id,
        platform=record.platform_version.platform.value,
        language=record.language,
        status=record.status.value,
        video_file_path=record.video_file_path,
        subtitle_file_path=record.subtitle_file_path,
        error_message=record.error_message,
        retry_of_id=record.retry_of_id,
        created_at=record.created_at,
        generated_at=record.generated_at,
        size_mb=file_size_mb(record.video_file_path),
    )


def run_generation_in_background(
    session_factory: sessionmaker,
    use_case: VideoGenerationUseCase,
    video_generation_id: str,
) -> None:
    """
    Execute one pending generation on its own session.

    Failures inside the pipeline are already recorded on the row. This also
    catches everything else, because an exception escaping a background task
    is swallowed by the event loop and would strand the row at PENDING with
    no explanation — a status the UI polls forever.
    """
    with session_factory() as session:
        try:
            use_case.run_pending(session, video_generation_id)
        except Exception as error:  # noqa: BLE001 - must not escape the task
            record = session.get(VideoGeneration, video_generation_id)
            if record is not None and record.status == VideoGenerationStatus.PENDING:
                record.status = VideoGenerationStatus.FAILED
                record.error_message = f"generation did not run: {error}"
                session.commit()


@router.post(
    "/chapters/{chapter_id}/video/generate",
    response_model=list[VideoGenerationResponse],
    status_code=202,
)
def generate_video(
    chapter_id: str,
    background_tasks: BackgroundTasks,
    request: Optional[VideoGenerateRequest] = None,
    session: Session = Depends(get_session),
    use_case: VideoGenerationUseCase = Depends(get_video_generation_use_case),
    session_factory: sessionmaker = Depends(get_session_factory),
) -> list[VideoGenerationResponse]:
    """
    Start generating videos; returns 202 immediately with pending rows.

    Composition runs in the background — poll GET /chapters/{id}/video for
    status. Defaults to every platform of the chapter in Spanish only;
    generating both languages doubles narration cost. Returns 409 when the
    script is unapproved or a pending/generated video already exists for that
    platform and language.
    """
    request = request or VideoGenerateRequest()
    chapter = _require_chapter(session, chapter_id)

    # Checked here rather than per row so a rejected request creates nothing.
    if not is_script_approved(chapter):
        raise HTTPException(
            status_code=409, detail="Script must be approved before video generation"
        )

    languages = list(request.languages or DEFAULT_LANGUAGES)
    for language in languages:
        if language not in SUPPORTED_LANGUAGES:
            raise HTTPException(
                status_code=422, detail=f"Unsupported language '{language}'."
            )

    platform_versions = _select_platform_versions(chapter, request.platforms)

    created = []
    for platform_version in platform_versions:
        for language in languages:
            try:
                created.append(
                    use_case.create_pending(session, platform_version.id, language)
                )
            except ScriptNotApprovedError as error:
                raise HTTPException(status_code=409, detail=str(error)) from error
            except GenerationAlreadyExistsError as error:
                # 409 rather than silently skipping: the caller asked for a
                # video that already exists, and should be told so rather
                # than shown a success that generated nothing.
                raise HTTPException(status_code=409, detail=str(error)) from error

    responses = [_to_response(record) for record in created]
    for record in created:
        background_tasks.add_task(
            run_generation_in_background, session_factory, use_case, record.id
        )
    return responses


@router.get("/chapters/{chapter_id}/video", response_model=list[VideoGenerationResponse])
def get_videos(
    chapter_id: str, session: Session = Depends(get_session)
) -> list[VideoGenerationResponse]:
    """Soft-deleted attempts are excluded here; GET /chapters/{id}/audit keeps them."""
    chapter = _require_chapter(session, chapter_id)
    return [
        _to_response(record)
        for platform_version in chapter.platform_versions
        for record in platform_version.video_generations
        if record.deleted_at is None
    ]


@router.post(
    "/chapters/{chapter_id}/video/retry",
    response_model=list[VideoGenerationResponse],
    status_code=202,
)
def retry_video(
    chapter_id: str,
    background_tasks: BackgroundTasks,
    request: Optional[VideoRetryRequest] = None,
    session: Session = Depends(get_session),
    use_case: VideoGenerationUseCase = Depends(get_video_generation_use_case),
    session_factory: sessionmaker = Depends(get_session_factory),
) -> list[VideoGenerationResponse]:
    """
    Retry failed generations, reusing audio and visuals already on disk.

    Retries every failed attempt for the chapter, or one when
    video_generation_id is given. Returns 409 when nothing has failed.
    """
    request = request or VideoRetryRequest()
    chapter = _require_chapter(session, chapter_id)

    failed = [
        record
        for platform_version in chapter.platform_versions
        for record in platform_version.video_generations
        # A deleted attempt is retired, not pending work to pick back up.
        if record.status == VideoGenerationStatus.FAILED and record.deleted_at is None
    ]
    if request.video_generation_id is not None:
        failed = [r for r in failed if r.id == request.video_generation_id]
        if not failed:
            raise HTTPException(
                status_code=404,
                detail=(
                    f"No failed VideoGeneration '{request.video_generation_id}' "
                    f"on chapter {chapter_id}."
                ),
            )

    if not failed:
        raise HTTPException(
            status_code=409, detail=f"Chapter {chapter_id} has no failed generations."
        )

    retries = []
    for record in failed:
        try:
            retries.append(use_case.create_retry(session, record.id))
        except ScriptNotApprovedError as error:
            raise HTTPException(status_code=409, detail=str(error)) from error

    responses = [_to_response(retry) for retry in retries]
    for retry in retries:
        background_tasks.add_task(
            run_generation_in_background, session_factory, use_case, retry.id
        )
    return responses


def _select_platform_versions(chapter: Chapter, platforms: Optional[list[str]]):
    if platforms is None:
        return list(chapter.platform_versions)

    selected = [
        platform_version
        for platform_version in chapter.platform_versions
        if platform_version.platform.value in platforms
    ]
    unknown = set(platforms) - {pv.platform.value for pv in chapter.platform_versions}
    if unknown:
        raise HTTPException(
            status_code=422,
            detail=f"Chapter {chapter.id} has no platform versions for: {sorted(unknown)}",
        )
    return selected


def _require_chapter(session: Session, chapter_id: str) -> Chapter:
    chapter = session.get(Chapter, chapter_id)
    if chapter is None:
        raise HTTPException(status_code=404, detail=f"Chapter {chapter_id} not found")
    return chapter